Move to Treesitter Main Branch
This has finally been completed now that the TextObjects main branch is
available

The new version of Treesitter produces a noticeable performance
improvement over the old one. Perhaps because it doesn't have to do an
autoload each time you enter insert mode. Because of this, the custom
.scm files have been moved to .bak files

The TextObjects config has more customized loading and more mappings as
well