Redo Statusline Based on mini.statusline
They key to this working was making sure to keep as much logic out of
the redraw process as possible, to minimize textlock, and to guard
against extra redraws.

Main benefit - Between the window management logic and the
function-based display, the event management is almost entirely gone. No more weird
little gaps between what is and what should be showing